Add services page and form

+++ database/migrations/2016_08_09_235447_create_services_table.php
@@ -19,7 +19,7 @@ class CreateServicesTable extends Migration
             $table->string('image');
             $table->boolean('active');
             $table->integer('sort');
-            $table->decimal('price');
+            $table->integer('price');
             $table->timestamps();
         });
     }
